refactoring of Game component
